bengaluru-house-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
